requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
django-extensions==3.2.3
python-decouple==3.8
mysqlclient==2.2.0
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
psutil==5.9.6
